move ServiceInfo to package com.ly.train.flower.registry.config
