+++ flower.common/src/main/java/com/ly/train/flower/registry/simple/SimpleRegistry.java
@@ -22,7 +22,7 @@ import java.util.List;
 import com.ly.train.flower.common.util.HttpClient;
 import com.ly.train.flower.common.util.URL;
 import com.ly.train.flower.registry.AbstractRegistry;
-import com.ly.train.flower.registry.ServiceInfo;
+import com.ly.train.flower.registry.config.ServiceInfo;
 
 /**
  * @author leeyazhou
