+++ flower.common/src/main/java/com/ly/train/flower/registry/zookeeper/ZookeeperRegistry.java
@@ -21,7 +21,7 @@ package com.ly.train.flower.registry.zookeeper;
 import java.util.List;
 import com.ly.train.flower.common.util.URL;
 import com.ly.train.flower.registry.Registry;
-import com.ly.train.flower.registry.ServiceInfo;
+import com.ly.train.flower.registry.config.ServiceInfo;
 
 /**
  * @author leeyazhou
